i2c-tools: add new tool 'i2ctransfer'
This tool allows to construct and concat multiple I2C messages into one
single transfer. Its aim is to test I2C master controllers, and so there
is no SMBus fallback.

I've been missing such a tool a number of times now, so I finally got
around to writing it myself. As with all I2C tools, it can be dangerous,
but it can also be very useful when developing.

It has been tested with various Renesas I2C IP cores as well as Tegra,
i.MX and AT91.

Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
Acked-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
[wsa: fixed two more typos and ugly linebreak]
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>